+++ NEWS.md
@@ -17,7 +17,7 @@
 * Export and document `tabnet_nn`
 * Added `min_grid.tabnet` method for `tune` (@cphaarmeyer, #107)
 * Added `tabnet_explain()` method for parsnip models (@cphaarmeyer, #108)
-* `tabnet_fit()` and `predict()` now allow **multi-outcome**, all numeric or all factors but not mixed. (#116)
+* `tabnet_fit()` and `predict()` now allow **multi-outcome**, all numeric or all factors but not mixed. (#118)
 
 ## Bugfixes